pyproject.toml
[project]
name = "outsera-challenge"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafael Ferreira", email = "perinzaum@proton.me"}
]
readme = "README.md"
requires-python = "3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
